backend/requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
pydantic==2.6.1
google-generativeai==0.7.2
chromadb==0.5.5
pymongo==4.8.0
requests==2.32.3
python-multipart==0.0.9
gTTS==2.5.3
SpeechRecognition==3.10.4
pydub==0.25.1
python-dotenv==1.0.1
